+++ src/app/(providers)/(root)/_components/DietsLog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DietsLog';
