+++ src/app/(providers)/(root)/_components/ExerciseTodoItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExerciseTodoItem';
