+++ src/app/(providers)/(root)/_components/ExerciseTodoList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExerciseTodoList';
