+++ src/app/(providers)/(root)/_components/LevelProgress/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LevelProgress';
