+++ src/app/(providers)/(root)/_components/TodoProgress/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TodoProgress';
