+++ src/app/(providers)/(root)/_components/WeightChart/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WeightChart';
